requirements.txt
aiofiles==0.4.0
aiohttp==3.5.1
aioredis==1.2.0
astroid==2.1.0
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
chardet==3.0.4
hiredis==0.3.1
httptools==0.0.11
idna==2.8
idna-ssl==1.1.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==5.0.0
multidict==4.5.2
pendulum==2.0.4
pluggy==0.8.0
py==1.7.0
pylint==2.2.2
pytest==4.1.0
pytest-aiohttp==0.3.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python-engineio==3.2.0
python-socketio==3.1.0
pytzdata==2018.7
redis==3.0.1
six==1.12.0
typed-ast==1.1.1
typing-extensions==3.6.6
ujson==1.35
urllib3==1.24.1
uvloop==0.11.3
websockets==6.0
wrapt==1.10.11
yarl==1.3.0
